+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Sorting by color key removed from formula category settings
 - Making ingredient category modal static
 - Making formula category modal static
+- Making ingredients advanced search modal static
 
 ### Version 7.4
 - Formula making view confirmation dialogs are now static